Set styles for display of title for one slide
We added a "title" image slide to the slideshow with text only in in the image. Updated styles for slideshow title element so that for this one slide, could have title text for slide display on vertical phone displays, where images are hidden. All larger displays show the 'title" image slide instead.